uni-56832: don't register tp_traverse ; seems to avoid the crash
There's a crash if we have a domain reload that wipes out the pythonnet assembly:
1. Some objects get leaked. (That's a bug in its own right.)
2. Leaked objects are on the gc list of garbage, even after a Py_Finalize (That's also a bug in its own right.)
3. Objects on the gc list get their tp_traverse function called during a gc collection
4. tp_traverse is a C# method
5. But we did a domain reload, so the C# method is gone... we're calling into garbage!

Solution: "if it hurts, don't do that." Just don't define tp_traverse, which
anyway is just the default implementation in both cases where it was defined.
(Also, two other related functions, also just default implementations.)

If we end up actually defining tp_traverse to do something we will have to
solve this for real.

Would be nice to limit the leaks also, and to improve the Py_Finalize/Py_Initialize loop.