Fix in dotnet code path: find mmap in libc rather than __Internal
When run with mono or in visual studio, DllImport("__Internal", ...) works; but
when run via dotnet, it fails to load the library somehow. mmap is in libc, so
get it that way now.

Exceptions thrown in NUnit setup/teardown get silently swallowed, so I also
added try/catch to dynamic.cs to help debug future initialization exceptions.